slide indexes for filtered/synced sliders
https://github.com/kenwheeler/slick/issues/515
https://github.com/kenwheeler/slick/issues/2500

The index passed into asNavFor is the original, data-slick- index, not the current, filtered index.

This change checks if filtering is happening, then looks at the target slides to see which has original index we are looking for. 
If it finds one, it corrects the index to hit that slide.

This can be tested on the demo page by filtering the sync demo: $('.slider-for, .slider-nav').slick('slickFilter',':even'), then clicking the 3 or 5.